requirements.txt
anthropic>=0.40.0
alpaca-py>=0.33.0
requests>=2.31.0
yfinance>=0.2.40
pandas>=2.0.0
python-dotenv>=1.0.0
